services/orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestrator"
version = "0.1.0"
description = "LangGraph-based agent orchestration service for Manus"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
